requirements.txt
openai
pandas
slack_sdk
google-ads==27.0.0
redis
streamlit
bs4
plotly
psycopg2-binary
numpy